fix bug integrate 1

+++ Final_Aptech_Sem4_Movie/src/main/java/com/demo/repositories/admin/PriceRepository.java
@@ -19,7 +19,7 @@ public interface PriceRepository extends PagingAndSortingRepository<Price, Integ
 	public Price findByDescription(@Param("day") String day);
 	
 	@Query("from Price where id = :id")
-	public Price findById(@Param("id") int id);
+	public Price findById(@Param("id") long id);
 
 }
